+++ General/Proc.vb
@@ -245,7 +245,10 @@ Public Class Proc
 
         Try
             If Header <> "" Then
-                If Not Package Is Nothing Then Header += " using " + Package.Name + " " + Package.Version
+                If Not Package Is Nothing Then
+                    Header += " using " + Package.Name + " " + Package.Version
+                End If
+
                 Log.WriteHeader(Header)
             End If