+++ Changelog.md
@@ -1,7 +1,9 @@
 
 ### 2.0.9.13 Beta not yet released
 
-
+- new: when a Event Command executes it writes a log entry,
+       this is now disabled by default but there is a new setting:
+       'Write Event Commands to log file'
 
 ### 2.0.9.12 Beta without apps